+++ async_upnp_client/profiles/profile.py
@@ -5,7 +5,6 @@ import asyncio
 import logging
 import time
 from datetime import timedelta
-from ipaddress import IPv4Address
 from typing import Any, Dict, FrozenSet, List, Optional, Sequence, Set, Union
 
 from async_upnp_client.client import (
@@ -16,7 +15,7 @@ from async_upnp_client.client import (
     UpnpStateVariable,
     UpnpValueError,
 )
-from async_upnp_client.const import SsdpHeaders
+from async_upnp_client.const import AddressTupleVXType, IPvXAddress, SsdpHeaders
 from async_upnp_client.event_handler import UpnpEventHandler
 from async_upnp_client.exceptions import (
     UpnpConnectionError,
@@ -59,14 +58,16 @@ class UpnpProfileDevice:
 
     @classmethod
     async def async_search(
-        cls, source_ip: Optional[IPv4Address] = None, timeout: int = SSDP_MX
+        cls,
+        source: Union[IPvXAddress, AddressTupleVXType, None] = None,
+        timeout: int = SSDP_MX,
     ) -> Set[SsdpHeaders]:
         """
         Search for this device type.
 
         This only returns search info, not a profile itself.
 
-        :param source_ip Source IP to scan from
+        :param source Source IP to scan from
         :param timeout Timeout to use
         :return: Set of devices (dicts) found
         """
@@ -76,9 +77,7 @@ class UpnpProfileDevice:
             if "st" in data and data["st"] in cls.DEVICE_TYPES:
                 responses.add(data)
 
-        await async_search(
-            async_callback=on_response, source_ip=source_ip, timeout=timeout
-        )
+        await async_search(async_callback=on_response, source=source, timeout=timeout)
 
         return responses