+++ async_upnp_client/ssdp_listener.py
@@ -4,8 +4,7 @@ import logging
 import re
 from asyncio.events import AbstractEventLoop
 from datetime import datetime, timedelta
-from ipaddress import ip_address
-from typing import Any, Awaitable, Callable, Mapping, Optional, Tuple
+from typing import Any, Awaitable, Callable, Mapping, Optional, Tuple, Union
 
 from async_upnp_client.advertisement import SsdpAdvertisementListener
 from async_upnp_client.const import (
@@ -20,13 +19,7 @@ from async_upnp_client.const import (
     UniqueDeviceName,
 )
 from async_upnp_client.search import SsdpSearchListener
-from async_upnp_client.ssdp import (
-    SSDP_IP_V4,
-    SSDP_IP_V6,
-    SSDP_MX,
-    SSDP_PORT,
-    udn_from_headers,
-)
+from async_upnp_client.ssdp import SSDP_MX, udn_from_headers
 from async_upnp_client.utils import CaseInsensitiveDict
 
 _LOGGER = logging.getLogger(__name__)
@@ -366,19 +359,16 @@ class SsdpListener:
         async_callback: Callable[
             [SsdpDevice, DeviceOrServiceType, SsdpSource], Awaitable
         ],
-        source_ip: Optional[IPvXAddress] = None,
-        target: Optional[AddressTupleVXType] = None,
+        source: Union[AddressTupleVXType, IPvXAddress, None] = None,
+        target: Union[AddressTupleVXType, IPvXAddress, None] = None,
         loop: Optional[AbstractEventLoop] = None,
         search_timeout: int = SSDP_MX,
     ) -> None:
         """Initialize."""
         # pylint: disable=too-many-arguments
         self.async_callback = async_callback
-        self.source_ip = source_ip
-        self.target: AddressTupleVXType = target or (
-            SSDP_IP_V4 if source_ip and source_ip.version == 4 else SSDP_IP_V6,
-            SSDP_PORT,
-        )
+        self.source = source
+        self.target = target
         self.loop = loop
         self.search_timeout = search_timeout
         self._device_tracker = SsdpDeviceTracker()
@@ -387,13 +377,12 @@ class SsdpListener:
 
     async def async_start(self) -> None:
         """Start search listener/advertisement listener."""
-        target_ip = ip_address(self.target[0])
         self._advertisement_listener = SsdpAdvertisementListener(
             on_alive=self._on_alive,
             on_update=self._on_update,
             on_byebye=self._on_byebye,
-            source_ip=self.source_ip,
-            target_ip=target_ip,
+            source=self.source,
+            target=self.target,
             loop=self.loop,
         )
         await self._advertisement_listener.async_start()
@@ -401,7 +390,7 @@ class SsdpListener:
         self._search_listener = SsdpSearchListener(
             self._on_search,
             loop=self.loop,
-            source_ip=self.source_ip,
+            source=self.source,
             target=self.target,
             timeout=self.search_timeout,
         )
@@ -478,3 +467,7 @@ class SsdpListener:
     def devices(self) -> Mapping[str, SsdpDevice]:
         """Get the known devices."""
         return self._device_tracker.devices
+
+    def __repr__(self) -> str:
+        """Return the string representation."""
+        return f"<{type(self).__name__}({self.source}, {self.target})>"