+++ CHANGES.rst
@@ -1,13 +1,17 @@
 Changes
 =======
 
-0.22.12 (unreleased)
+0.23.0 (unreleased)
+
+- Fix error in cli if target_port is absent in search command
+- Fix IPv6 support for SSDP
+    - Change `ssdp_listener`, `advertisement`, `search` to use `AddressTupleVXType` or `IPvXAddress`
+    - Support for Python <3.9, use AddressTupleV6Type when scope_ids are needed
 
 
 0.22.11 (2021-10-31)
 
 - Poll state variables when event subscriptions are rejected (@chishm)
-- Fix error in cli if target_port is absent in search command
 
 
 0.22.10 (2021-10-25)